Street Fighter II - Champion Edition - READ NOTE!
Any Clones that use this Parent Rom, will use the same Codes. Will have to create for each and every one of them unfortunately. So if I miss one that you have.....let me know in the Issues Section so I can add it. But as far as I can tell, all the Clones use the same Codes as the Parent Roms.......so yeah.